backend/bot/requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
aiomqtt==2.1.0
